Added usual "pyplot environment setup" functions.
This makes it possible to keep the enhancements brought by the
rc_context manager. Indeed, the Axes object need to be created within
the right rc_context.

These functions should be used in replacement of the usual plt.subplots
and plt.subplots2grid functions to build the Axes later sent to ppl
plotting functions.